requirements.txt
 # Core
fastapi
uvicorn
streamlit
boto3
pydantic
python-multipart

# LLM & Embedding
langchain
sentence-transformers
faiss-cpu

# PDF Parsing
PyMuPDF

# CORS
fastapi[all]

